requirements.txt
numpy
opencv-python
pdf2image
Pillow